tg-bot/bot/requirements.txt
aiogram>=3.0
requests>=2.28.0
python-dotenv>=0.19.0
